Amended the test script command and renamed the controller and model
