+++ .travis.yml
@@ -1,6 +1,3 @@
 language: node_js
 node_js:
   - "6.9.1"
-
-script:
-  - npm run ci
